requirements.txt
# LLM and Orchestration
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5
anthropic>=0.18.0
openai>=1.12.0

# Vector Database
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.3.1

# Data Validation
pydantic>=2.5.0

# Environment Management
python-dotenv>=1.0.0

# Text Processing
beautifulsoup4>=4.12.0
spacy>=3.7.0
lxml>=5.1.0

# Data Manipulation
pandas>=2.1.0
numpy>=1.26.0

# CLI
click>=8.1.7
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
